+++ packages/messaging/src/events/post/payloads.ts
@@ -1,13 +1,13 @@
+import type { IEventIdPayload, IPostIdPayload, IUserIdPayload } from '../index.js';
+
 export enum PostEventMessagingType {
   POST_CREATED = 'post.created',
   POST_DELETED = 'post.deleted',
 }
 
-export interface IPostCreatedPayload {
-  id: string;
-  eventId?: string;
-}
+export interface IPostCreatedPayload
+  extends IPostIdPayload,
+    Partial<IEventIdPayload>,
+    Partial<IUserIdPayload> {}
 
-export interface IPostDeletedPayload {
-  id: string;
-}
+export interface IPostDeletedPayload extends IPostIdPayload, Partial<IUserIdPayload> {}