+++ packages/messaging/src/websockets/users/payload.ts
@@ -1,0 +1,16 @@
+import type {
+  IEmitterPayload,
+  IUserSocialCreatedPayload,
+  IUserSocialCreationFailedPayload,
+  IUserSocialDeletedPayload,
+  IUserSocialDeletionFailedPayload,
+} from '../../index.js';
+
+export interface IUserCreatedWebsocketPayload extends IUserSocialCreatedPayload, IEmitterPayload {}
+export interface IUserCreationFailedWebsocketPayload
+  extends IUserSocialCreationFailedPayload,
+    IEmitterPayload {}
+export interface IUserDeletedWebsocketPayload extends IUserSocialDeletedPayload, IEmitterPayload {}
+export interface IUserDeletionFailedWebsocketPayload
+  extends IUserSocialDeletionFailedPayload,
+    IEmitterPayload {}