+++ packages/messaging/src/websockets/users/types.ts
@@ -1,0 +1,6 @@
+export enum UserWebsocketMessagingType {
+  USER_CREATED = 'user.created',
+  USER_DELETED = 'user.deleted',
+  USER_CREATION_FAILED = 'user.creation_failed',
+  USER_DELETION_FAILED = 'user.deletion_failed',
+}