+++ packages/shared/src/enums/streams.enum.ts
@@ -1,18 +1,18 @@
 export enum UserStream {
   USER_JOB_OUTBOX_SUCCESS = 'user:job:outbox:success',
   USER_JOB_OUTBOX_FAILURE = 'user:job:outbox:failure',
-  USER_USERS = 'user:users',
-  USER_AUTHORIZATIONS = 'user:authorizations',
-  USER_BADGES = 'user:badges',
-  USER_NOTIFICATIONS = 'user:notifications',
+  USER_CREATED = 'user:created',
+  USER_DELETED = 'user:deleted',
 }
 
 export enum EventStream {
   EVENT_JOB_OUTBOX_SUCCESS = 'event:job:outbox:success',
   EVENT_JOB_OUTBOX_FAILURE = 'event:job:outbox:failure',
-  EVENT_EVENTS = 'event:events',
+  EVENT_CREATED = 'event:created',
+  EVENT_DELETED = 'event:deleted',
   EVENT_TAGS = 'event:tags',
   EVENT_REQUIREMENTS = 'event:requirements',
+  EVENT_SUCCESSFULLY_CREATED = 'event:successfully_created',
 }
 
 export enum PostStream {
@@ -34,12 +34,12 @@ export enum SocialStream {
 }
 
 export enum WebsocketStream {
-  WS_USER = 'ws:user',
-  WS_EVENT = 'ws:event',
-  WS_TAG = 'ws:tag',
-  WS_BADGE = 'ws:badge',
+  WS_USER_CREATED_FEEDBACK = 'ws:user-created-feedback',
+  WS_USER_DELETED_FEEDBACK = 'ws:user-deleted-feedback',
   WS_POST_CREATED_FEEDBACK = 'ws:post-created-feedback',
   WS_POST_DELETED_FEEDBACK = 'ws:post-deleted-feedback',
+  WS_EVENT_CREATED_FEEDBACK = 'ws:event-created-feedback',
+  WS_EVENT_DELETED_FEEDBACK = 'ws:event-deleted-feedback',
 }
 
 export const Streams = {