+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.3.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/shared@0.7.1
+
 ## 3.3.5
 
 ### Patch Changes