+++ packages/domain-post/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.5.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.8.2
+  - @volontariapp/outbox@0.9.31
+  - @volontariapp/database@3.3.13
+
 ## 3.5.2
 
 ### Patch Changes